added num_workers, cluster visualization, benchmarked time

+++ utils.py
@@ -22,7 +22,7 @@ def show_images(images, cols = 5, titles = None):
     if titles is None: titles = ['Image (%d)' % i for i in range(1,n_images + 1)]
     fig = plt.figure()
     for n, (image, title) in enumerate(zip(images, titles)):
-        a = fig.add_subplot(cols, np.ceil(n_images/float(cols)), n + 1)
+        a = fig.add_subplot(cols, int(np.ceil(n_images/float(cols))), n + 1)
         if image.ndim == 2:
             plt.gray()
         plt.imshow(image)
